feat: Add components method to Variant class
Adds a `components` method to the Variant class which is a tuple
of fully qualified variant strings that comprises the current variant
object. The `format_variant` function was refactored into a new
function `_format_component_variants`, which returns a list of the
component variants of a Variant object. In turn, the __repr__ function
uses this class function to return the variant string representation.
The `components` method returns the result of this new internal
formatting function, with the prefix and transcript identifier prepended
to each variant string.

For multi-variants, each tuple element is parsable into a single-variant
Variant object which comprises the original multi-variant. For single-
variants, the single element tuple contains the original variant string.